+++ src/handlers/draw-figures/draw-circle.ts
