+++ src/handlers/draw-figures/draw-rectangle.ts
