+++ src/handlers/draw-figures/draw-square.ts
