+++ src/handlers/draw-figures/index.ts
