+++ src/handlers/types.ts
@@ -1,0 +1,4 @@
+export type MOUSE_MOVE_COMMANDS = 'mouse_up' | 'mouse_down' | 'mouse_left' | 'mouse_right';
+export type DROW_FIGURES_COMMANDS = 'draw_circle' | 'draw_rectangle' | 'draw_square';
+
+export type COMMANDS = MOUSE_MOVE_COMMANDS | DROW_FIGURES_COMMANDS | 'prnt_scrn' | 'mouse_position'
