--- src/mouse-handlers/types.ts
@@ -1,12 +1,0 @@
-export type COMMANDS = 
-	'mouse_up' | 
-	'mouse_down' |
-	'mouse_left' |
-	'mouse_right' |
-	'mouse_position' |
-	'draw_circle' |
-	'draw_rectangle' |
-	'draw_square' |
-	'prnt_scrn';
-	
-	//TODO разделить на категории
